Refactoring: Moved table to separate component
