+++ src/App.tsx
@@ -1,11 +1,14 @@
-
-import { GeneralTime } from './GeneralTime';
+import { GeneralTime } from './components/GeneralTime.tsx';
+import {Table} from "./components/Table.tsx";
 
 
 export const App = ()=> {
 
-  return <>
-  <GeneralTime />
-  </>
+  return (
+      <>
+        <GeneralTime/>
+        <Table/>
+      </>
+  )
 }
